refactor: move failover logic to failover connection plugin
