+++ src/main/resources/com/mysql/cj/LocalizedErrorMessages.properties
@@ -998,6 +998,7 @@ PropertyDefinition.1=The connection property ''{0}'' acceptable values are: {1}.
 
 ConnectionProperties.clusterId=A unique identifier for the cluster. Connections with the same cluster id share a cluster topology cache. If unspecified, a cluster id is automatically created for AWS RDS clusters.
 ConnectionProperties.clusterInstanceHostPattern=The cluster instance DNS pattern that will be used to build a complete instance endpoint. A "?" character in this pattern should be used as a placeholder for cluster instance names. This pattern is required to be specified for IP address or custom domain connections to AWS RDS clusters. Otherwise, if unspecified, the pattern will be automatically created for AWS RDS clusters.
+ConnectionProperties.useConnectionPlugins=Use connection plugins. 
 ConnectionProperties.enableClusterAwareFailover=Enable/disable cluster-aware failover logic. 
 ConnectionProperties.clusterTopologyRefreshRateMs=Cluster topology refresh rate in millis. The cached topology for the cluster will be invalidated after the specified time, after which it will be updated during the next interaction with the connection.
 ConnectionProperties.failoverTimeoutMs=Maximum allowed time in millis to attempt reconnecting to a new writer or reader instance after a cluster failover is initiated.
