--- src/main/user-impl/java/com/mysql/cj/jdbc/ha/ca/ReaderFailoverHandler.java
@@ -1,63 +1,0 @@
-/*
- * AWS JDBC Driver for MySQL
- * Copyright Amazon.com Inc. or affiliates.
- *
- * This program is free software; you can redistribute it and/or modify it under
- * the terms of the GNU General Public License, version 2.0, as published by the
- * Free Software Foundation.
- *
- * This program is also distributed with certain software (including but not
- * limited to OpenSSL) that is licensed under separate terms, as designated in a
- * particular file or component or in included license documentation. The
- * authors of this program hereby grant you an additional permission to link the
- * program and your derivative works with the separately licensed software that
- * they have included with MySQL.
- *
- * Without limiting anything contained in the foregoing, this file, which is
- * part of this connector, is also subject to the Universal FOSS Exception,
- * version 1.0, a copy of which can be found at
- * http://oss.oracle.com/licenses/universal-foss-exception.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU General Public License, version 2.0,
- * for more details.
- *
- * You should have received a copy of the GNU General Public License along with
- * this program; if not, write to the Free Software Foundation, Inc.,
- * 51 Franklin St, Fifth Floor, Boston, MA 02110-1301  USA
- */
-
-package com.mysql.cj.jdbc.ha.ca;
-
-import com.mysql.cj.conf.HostInfo;
-
-import java.sql.SQLException;
-import java.util.List;
-
-/**
- * Interface for Reader Failover Process handler. This handler implements all necessary logic to try
- * to reconnect to another reader host.
- */
-public interface ReaderFailoverHandler {
-
-  /**
-   * Called to start Reader Failover Process. This process tries to connect to any reader. If no
-   * reader is available then driver may also try to connect to a writer host, down hosts, and the
-   * current reader host.
-   *
-   * @param hosts Cluster current topology.
-   * @param currentHost The currently connected host that has failed.
-   * @return {@link ReaderFailoverResult} The results of this process.
-   */
-  ReaderFailoverResult failover(List<HostInfo> hosts, HostInfo currentHost) throws SQLException;
-
-  /**
-   * Called to get any available reader connection. If no reader is available then result of process
-   * is unsuccessful. This process will not attempt to connect to the writer host.
-   *
-   * @param hostList Cluster current topology.
-   * @return {@link ReaderFailoverResult} The results of this process.
-   */
-  ReaderFailoverResult getReaderConnection(List<HostInfo> hostList) throws SQLException;
-}
