+++ src/main/user-impl/java/com/mysql/cj/jdbc/ha/ca/plugins/failover/ClusterAwareReaderFailoverHandler.java
@@ -2,37 +2,37 @@
  * AWS JDBC Driver for MySQL
  * Copyright Amazon.com Inc. or affiliates.
  *
- * This program is free software; you can redistribute it and/or modify it under
- * the terms of the GNU General Public License, version 2.0, as published by the
- * Free Software Foundation.
+ * Redistribution and use in source and binary forms, with or without modification,
+ * are permitted provided that the following conditions are met:
  *
- * This program is also distributed with certain software (including but not
- * limited to OpenSSL) that is licensed under separate terms, as designated in a
- * particular file or component or in included license documentation. The
- * authors of this program hereby grant you an additional permission to link the
- * program and your derivative works with the separately licensed software that
- * they have included with MySQL.
+ * 1. Redistributions of source code must retain the above copyright notice,
+ * this list of conditions and the following disclaimer.
  *
- * Without limiting anything contained in the foregoing, this file, which is
- * part of this connector, is also subject to the Universal FOSS Exception,
- * version 1.0, a copy of which can be found at
- * http://oss.oracle.com/licenses/universal-foss-exception.
+ * 2. Redistributions in binary form must reproduce the above copyright notice,
+ * this list of conditions and the following disclaimer in the documentation and/or
+ * other materials provided with the distribution.
  *
- * This program is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU General Public License, version 2.0,
- * for more details.
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY
+ * EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES
+ * OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT
+ * SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
+ * INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
+ * PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
+ * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
+ * LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  *
- * You should have received a copy of the GNU General Public License along with
- * this program; if not, write to the Free Software Foundation, Inc.,
- * 51 Franklin St, Fifth Floor, Boston, MA 02110-1301  USA
  */
 
-package com.mysql.cj.jdbc.ha.ca;
+package com.mysql.cj.jdbc.ha.ca.plugins.failover;
 
 import com.mysql.cj.Messages;
 import com.mysql.cj.conf.HostInfo;
 import com.mysql.cj.jdbc.JdbcConnection;
+import com.mysql.cj.jdbc.ha.ca.ClusterAwareConnectionProxy;
+import com.mysql.cj.jdbc.ha.ca.ClusterAwareUtils;
+import com.mysql.cj.jdbc.ha.ca.ConnectionProvider;
+import com.mysql.cj.jdbc.ha.ca.TopologyService;
 import com.mysql.cj.log.Log;
 import com.mysql.cj.log.NullLogger;
 import com.mysql.cj.util.Util;
@@ -158,7 +158,7 @@ public class ClusterAwareReaderFailoverHandler implements ReaderFailoverHandler 
       this.log.logDebug(Messages.getString("ClusterAwareReaderFailover.6", new Object[] {"failover"}));
       return new ReaderFailoverResult(
           null,
-          ClusterAwareConnectionProxy.NO_CONNECTION_INDEX,
+          FailoverConnectionPlugin.WRITER_CONNECTION_INDEX,
           false);
     }
 
@@ -188,7 +188,7 @@ public class ClusterAwareReaderFailoverHandler implements ReaderFailoverHandler 
       ExecutorService executor,
       Future<ReaderFailoverResult> future) throws SQLException {
     ReaderFailoverResult defaultResult = new ReaderFailoverResult(
-        null, ClusterAwareConnectionProxy.NO_CONNECTION_INDEX, false);
+        null, FailoverConnectionPlugin.WRITER_CONNECTION_INDEX, false);
     try {
       ReaderFailoverResult result =
           future.get(this.maxFailoverTimeoutMs, TimeUnit.MILLISECONDS);
@@ -223,14 +223,14 @@ public class ClusterAwareReaderFailoverHandler implements ReaderFailoverHandler 
     return getConnectionFromHostGroup(hostGroup);
   }
 
-  List<HostTuple> getHostTuplesByPriority(List<HostInfo> hosts, Set<String> downHosts) {
+  public List<HostTuple> getHostTuplesByPriority(List<HostInfo> hosts, Set<String> downHosts) {
     List<HostTuple> hostGroup = new ArrayList<>();
     addActiveReaders(hostGroup, hosts, downHosts);
-    HostInfo writerHost = hosts.get(ClusterAwareConnectionProxy.WRITER_CONNECTION_INDEX);
+    HostInfo writerHost = hosts.get(FailoverConnectionPlugin.WRITER_CONNECTION_INDEX);
     hostGroup.add(
         new HostTuple(
             writerHost,
-            ClusterAwareConnectionProxy.WRITER_CONNECTION_INDEX));
+            FailoverConnectionPlugin.WRITER_CONNECTION_INDEX));
     addDownHosts(hostGroup, hosts, downHosts);
     return hostGroup;
   }
@@ -240,7 +240,7 @@ public class ClusterAwareReaderFailoverHandler implements ReaderFailoverHandler 
       List<HostInfo> hosts,
       Set<String> downHosts) {
     List<HostTuple> activeReaders = new ArrayList<>();
-    for (int i = ClusterAwareConnectionProxy.WRITER_CONNECTION_INDEX + 1;
+    for (int i = FailoverConnectionPlugin.WRITER_CONNECTION_INDEX + 1;
          i < hosts.size(); i++) {
       HostInfo host = hosts.get(i);
       if (!downHosts.contains(host.getHostPortPair())) {
@@ -280,7 +280,7 @@ public class ClusterAwareReaderFailoverHandler implements ReaderFailoverHandler 
       this.log.logDebug(Messages.getString("ClusterAwareReaderFailover.6", new Object[] {"getReaderConnection"}));
       return new ReaderFailoverResult(
           null,
-          ClusterAwareConnectionProxy.NO_CONNECTION_INDEX,
+          FailoverConnectionPlugin.WRITER_CONNECTION_INDEX,
           false);
     }
 
@@ -289,7 +289,7 @@ public class ClusterAwareReaderFailoverHandler implements ReaderFailoverHandler 
     return getConnectionFromHostGroup(tuples);
   }
 
-  List<HostTuple> getReaderTuplesByPriority(
+  public List<HostTuple> getReaderTuplesByPriority(
       List<HostInfo> hostList,
       Set<String> downHosts) {
     List<HostTuple> tuples = new ArrayList<>();
@@ -303,7 +303,7 @@ public class ClusterAwareReaderFailoverHandler implements ReaderFailoverHandler 
       List<HostInfo> hosts,
       Set<String> downHosts) {
     List<HostTuple> downReaders = new ArrayList<>();
-    for (int i = ClusterAwareConnectionProxy.WRITER_CONNECTION_INDEX + 1; i < hosts.size(); i++) {
+    for (int i = FailoverConnectionPlugin.WRITER_CONNECTION_INDEX + 1; i < hosts.size(); i++) {
       HostInfo host = hosts.get(i);
       if (downHosts.contains(host.getHostPortPair())) {
         downReaders.add(new HostTuple(host, i));
@@ -339,7 +339,7 @@ public class ClusterAwareReaderFailoverHandler implements ReaderFailoverHandler 
 
       return new ReaderFailoverResult(
           null,
-          ClusterAwareConnectionProxy.NO_CONNECTION_INDEX,
+          FailoverConnectionPlugin.WRITER_CONNECTION_INDEX,
           false);
     } finally {
       executor.shutdownNow();
@@ -370,14 +370,14 @@ public class ClusterAwareReaderFailoverHandler implements ReaderFailoverHandler 
     }
     return new ReaderFailoverResult(
         null,
-        ClusterAwareConnectionProxy.NO_CONNECTION_INDEX,
+        FailoverConnectionPlugin.WRITER_CONNECTION_INDEX,
         false);
   }
 
   private ReaderFailoverResult getNextResult(CompletionService<ReaderFailoverResult> service)
       throws SQLException {
     ReaderFailoverResult defaultResult = new ReaderFailoverResult(
-        null, ClusterAwareConnectionProxy.NO_CONNECTION_INDEX, false);
+        null, FailoverConnectionPlugin.WRITER_CONNECTION_INDEX, false);
     try {
       Future<ReaderFailoverResult> future =
           service.poll(this.timeoutMs, TimeUnit.MILLISECONDS);
@@ -434,7 +434,7 @@ public class ClusterAwareReaderFailoverHandler implements ReaderFailoverHandler 
                 new Object[] {this.newHostTuple.getIndex(), newHost.getHostPortPair()}));
         return new ReaderFailoverResult(
             null,
-            ClusterAwareConnectionProxy.NO_CONNECTION_INDEX,
+            FailoverConnectionPlugin.WRITER_CONNECTION_INDEX,
             false);
       }
     }
@@ -443,11 +443,11 @@ public class ClusterAwareReaderFailoverHandler implements ReaderFailoverHandler 
   /**
    * HostTuple class.
    * */
-  protected static class HostTuple {
+  public static class HostTuple {
     private final HostInfo host;
     private final int index;
 
-    HostTuple(HostInfo host, int index) {
+    public HostTuple(HostInfo host, int index) {
       this.host = host;
       this.index = index;
     }
