+++ src/test/java/com/mysql/cj/jdbc/ha/ca/ClusterAwareWriterFailoverHandlerTest.java
@@ -33,6 +33,10 @@ package com.mysql.cj.jdbc.ha.ca;
 import com.mysql.cj.conf.HostInfo;
 import com.mysql.cj.jdbc.ConnectionImpl;
 import com.mysql.cj.jdbc.JdbcConnection;
+import com.mysql.cj.jdbc.ha.ca.plugins.failover.ClusterAwareWriterFailoverHandler;
+import com.mysql.cj.jdbc.ha.ca.plugins.failover.ReaderFailoverHandler;
+import com.mysql.cj.jdbc.ha.ca.plugins.failover.ReaderFailoverResult;
+import com.mysql.cj.jdbc.ha.ca.plugins.failover.WriterFailoverResult;
 import com.mysql.cj.log.Log;
 import org.junit.jupiter.api.Test;
 import org.mockito.ArgumentMatchers;
